refactor: AdminController 통합, 관리자 로그인 메인 페이지 연결
