+++ backend/src/main/java/kr/co/programmers/cafe/domain/admin/AdminController.java
@@ -4,9 +4,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Slf4j
 @Controller
@@ -35,5 +36,25 @@ public class AdminController {
     public String adminManage() {
         log.info("관리 페이지 호출 테스트 - 로그인 성공");
         return "admin/manage-form";
+    }
+
+    @GetMapping("/items")
+    public String items(Model model) {
+        /*Item item = Item.builder().name("상품1").price(1000).id(1L).category(Item.Category.Coffee).imageUrl("temp.png").build();
+        List<Item> items = new ArrayList<>();
+        items.add(item);
+        model.addAttribute("items", items);*/
+        return "items/item-list";
+    }
+
+    @GetMapping
+    public String main(){
+        return "main";
+    }
+
+    @DeleteMapping("/items/{itemId}")
+    public String deleteItems(@PathVariable Long itemId){
+        log.info("itemId={} Deleted", itemId);
+        return "redirect:/admin/items";
     }
 }
